requirements.txt
asgiref==3.6.0
bcrypt==3.2.0
boto3==1.21.19
botocore==1.24.19
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==15.3.4
idna==3.3
jmespath==0.10.0
linode-api4==5.2.1
oauthlib==3.2.2
openpyxl==3.0.9
paramiko==2.10.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.8
